dashboard/requirements.txt
streamlit
requests
psycopg2
python-dotenv
